add: IProductDataSource interface
- add: products-memory-data-source
